u32 api_version to align with VERSION, and use VERSION instead of adding base_version

+++ substrate/primitives/api/proc-macro/src/impl_runtime_apis.rs
@@ -462,7 +462,7 @@ fn extend_with_runtime_decl_path(mut trait_: Path) -> Path {
 	trait_
 }
 
-fn extend_with_api_version(mut trait_: Path, version: Option<u64>) -> Path {
+fn extend_with_api_version(mut trait_: Path, version: Option<u32>) -> Path {
 	let version = if let Some(v) = version {
 		v
 	} else {
